working on opcodes

+++ src/stack.hpp
@@ -5,5 +5,5 @@
 namespace chip8
 {
 uint16_t stack[0x10];
-uint8_t stack_pointer;
+uint16_t stack_pointer;
 }
